requirements.txt
anytree
aria2p
asyncio
beautifulsoup4
bencoding
cfscrape
dnspython
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
psutil
pybase64
pymongo
pyrogram
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
telegraph
tenacity
tgCrypto
xattr
yt-dlp